refactor: 토큰 만료시간 테스트 추가 및 패키지 변경
